[Fix] Character Die
캐릭터 죽게 되고 다시 리스폰 했을 때 리스폰 딜레이 시간만큼 동작을 못하는 버그 수정
