+++ HIGHFIVE/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -39,19 +39,14 @@ MonoBehaviour:
   - DestroySpaceship
   - Fire
   - RespawnSpaceship
-  - SyncReadyStatus
-  - SetReadyImage
-  - SetLayer
-  - ToTarget
-  - SetTarget
-  - Destroy
-  - SetHpRatio
-  - SyncHpRatio
-  - RPCAddPhotonView
   - RPCDetroy
-  - TakeDamage
-  - SetHpRPC
   - SetHpBarColor
+  - SetHpRPC
+  - SetLayer
+  - SetReadyImage
+  - SetTarget
+  - SyncHpRatio
+  - ToTarget
   DisableAutoOpenWizard: 1
   ShowSettings: 1
   DevRegionSetOnce: 1
