+++ HIGHFIVE/Assets/Scripts/Object/Character/Character.cs
@@ -111,36 +111,4 @@ public class Character : Creature
         _playerStateMachine._player.Animator.SetBool(_playerStateMachine._player.PlayerAnimationData.DieParameterHash, false);
         Camera.main.transform.position = new Vector3(transform.position.x, transform.position.y, -10f);
     }
-
-
-    [PunRPC]
-    public void SetLayer(int layer)
-    {
-        gameObject.layer = layer;
-    }
-
-    [PunRPC]
-    public void SetHpBarColor()
-    {
-        Image fillImage = null;
-        foreach (Image component in gameObject.GetComponentsInChildren<Image>())
-        {
-            if (component.name == "Fill")
-            {
-                fillImage = component;
-            }
-        }
-        if (fillImage != null)
-        {
-            if (gameObject.layer == (int)Define.Camp.Red)
-            {
-                fillImage.color = Define.GreenColor;
-            }
-            else
-            {
-                fillImage.color = Define.BlueColor;
-            }
-
-        }
-    }
 }
