--- HIGHFIVE/Assets/Scripts/Photon/LayerSetter.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LayerSetter : MonoBehaviour
-{
-    private void Start()
-    {
-        
-    }
-
-    private void Update()
-    {
-        
-    }
-}
