+++ HIGHFIVE/Assets/Scripts/Scene/GameScene.cs
@@ -30,10 +30,10 @@ public class GameScene : BaseScene
         if (_classMapping.TryGetValue(Main.GameManager.SelectedCharacter, out selectClass))
         {
             characterObj = Main.ResourceManager.Instantiate($"Character/{selectClass}", position, syncRequired: true);
-            characterObj.layer = Main.GameManager.SelectedCamp == Define.Camp.Red ? (int)Define.Layer.Red : (int)Define.Layer.Blue;
+            int layer = Main.GameManager.SelectedCamp == Define.Camp.Red ? (int)Define.Layer.Red : (int)Define.Layer.Blue;
             Main.ResourceManager.Instantiate("UI_Prefabs/GameSceneUI");
             Main.GameManager.SpawnedCharacter = characterObj.GetComponent<Character>();
-            characterObj.GetComponent<PhotonView>().RPC("SetLayer", RpcTarget.All, characterObj.layer);
+            characterObj.GetComponent<PhotonView>().RPC("SetLayer", RpcTarget.All, layer);
             _cameraController.CallCharacterSpawnEvent();
         }
     }
